+++ tutorials/plot_stats_cluster_spatio_temporal_2samp.py
@@ -1,6 +1,4 @@
 """
-.. _tut_stats_cluster_source_2samp:
-
 =========================================================================
 2 samples permutation test on source data with spatio-temporal clustering
 =========================================================================