+++ tutorials/plot_stats_cluster_time_frequency_repeated_measures_anova.py
@@ -1,6 +1,4 @@
 """
-.. _tut_stats_cluster_sensor_rANOVA_tfr:
-
 ====================================================================
 Mass-univariate twoway repeated measures ANOVA on single trial power
 ====================================================================